requirements.txt
google-cloud-bigquery>=3.13.0
google-cloud-storage>=2.10.0
pyarrow>=14.0.1
pandas>=2.2.0
google-cloud-bigquery-connection>=1.12.0
fastavro>=1.8.0